requirements.txt
mysql-connector-python
openai
requests
python-dotenv
selenium
webdriver_manager
clipboard